Cleanup of unused stuff
