+++ package.json
@@ -61,7 +61,6 @@
     "hapi-decorators": "^2.0.0",
     "moment": "^2.24.0",
     "pg": "^8.2.1",
-    "ts-postgres": "^1.1.3",
     "typeorm": "^0.2.25",
     "uuid": "^8.1.0",
     "winston": "^3.2.1"
