--- src/interfaces/drawing-number.ts
@@ -1,6 +1,0 @@
-import { PowerballNumber } from './powerball-number';
-
-export interface DrawingNumber extends PowerballNumber {
-  ticketId: string;
-  powerPlay: boolean;
-}
