--- src/interfaces/powerball-ticket-purchase.ts
@@ -1,9 +1,0 @@
-import { PowerballTicket } from './powerball-ticket';
-import { PowerballTicketNumber } from './powerball-ticket-number';
-import { PowerballTicketDrawing } from './powerball-ticket-drawing';
-
-export interface PowerballTicketPurchase {
-  ticket: PowerballTicket;
-  numbers: Array<PowerballTicketNumber>;
-  drawings: Array<PowerballTicketDrawing>;
-}
