+++ src/interfaces/powerball-ticket-report.ts
@@ -1,6 +1,11 @@
 import { OwnerWinning } from './owner-winning';
-import { PowerballTicketPurchase } from './powerball-ticket-purchase';
+import { PowerballTicket } from './powerball-ticket';
+import { PowerballTicketNumber } from './powerball-ticket-number';
+import { PowerballTicketDrawing } from './powerball-ticket-drawing';
 
-export interface PowerballTicketReport extends PowerballTicketPurchase {
+export interface PowerballTicketReport {
+  ticket: PowerballTicket;
+  numbers: Array<PowerballTicketNumber>;
+  drawings: Array<PowerballTicketDrawing>;
   winnings: Array<OwnerWinning>;
 }
